move Tab component to its own file
