+++ packages/react-native/src/primitives/Tabs/index.ts
@@ -1,2 +1,3 @@
-export { Tab, Tabs } from './Tabs';
+export { default as Tab } from './Tab';
+export { default as Tabs } from './Tabs';
 export { TabProps, TabsProps } from './types';
